requirements.txt
torch>=2.0
numpy
matplotlib
pyyaml
transformers
accelerate
datasets
pandas
scipy
scikit-learn
umap-learn
